+++ tests/sdk/v2/driver/test_client.py
@@ -16,7 +16,7 @@ from orquestra.sdk._base._driver import _exceptions
 from orquestra.sdk._base._driver._client import DriverClient, Paginated
 from orquestra.sdk._base._driver._models import Resources
 from orquestra.sdk.schema.ir import WorkflowDef
-from orquestra.sdk.schema.responses import JSONResult, PickleResult, WorkflowResult
+from orquestra.sdk.schema.responses import JSONResult, PickleResult
 from orquestra.sdk.schema.workflow_run import RunStatus, State, TaskRun
 
 from . import resp_mocks
@@ -56,7 +56,7 @@ def endpoint_mocker_base(mocked_responses):
 class TestClient:
     @pytest.fixture
     def base_uri(self):
-        return "https://shouldnt.matter.example.com"
+        return "https://should.not.matter.example.com"
 
     @pytest.fixture
     def token(self):
@@ -1602,6 +1602,86 @@ class TestClient:
                 )
 
             @staticmethod
+            def test_logs_decode(
+                endpoint_mocker, client: DriverClient, workflow_run_id: str
+            ):
+                endpoint_mocker(
+                    body=resp_mocks.make_get_wf_run_logs_response_with_content(),
+                    match=[
+                        responses.matchers.query_param_matcher(
+                            {"workflowRunId": workflow_run_id}
+                        )
+                    ],
+                )
+
+                assert client.get_workflow_run_logs(workflow_run_id) == [
+                    ":actor_name:Manager",
+                    ":actor_name:Manager",
+                    ":task_name:create_ray_workflow",
+                    '2023-03-14 14:32:57,554\tINFO api.py:203 -- Workflow job created. [id="wf.wf_def.3c5f938"].',  # noqa: E501
+                    ":task_name:_workflow_task_executor_remote",
+                    "2023-03-14 14:32:57,933\tINFO task_executor.py:78 -- Task status [RUNNING]\t[wf.wf_def.3c5f938@invocation-0-task-do-thing]",  # noqa: E501
+                    'Traceback (most recent call last):\n  File "/home/orquestra/venv/lib/python3.9/site-packages/orquestra/sdk/_ray/_dag.py", line 192, in _ray_remote\n    return wrapped(*inner_args, **inner_kwargs)\n  File "/home/orquestra/venv/lib/python3.9/site-packages/orquestra/sdk/_ray/_dag.py", line 147, in __call__\n    return self._fn(*unpacked_args, **unpacked_kwargs)\n  File "/Users/benjaminmummery/Documents/Projects/orquestra-workflow-sdk/../scratch/scratch.py", line 12, in do_thing\nException: Blah\n',  # noqa: E501
+                    "Traceback (most recent call last):",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/orquestra/sdk/_ray/_dag.py", line 192, in _ray_remote',  # noqa: E501
+                    "    return wrapped(*inner_args, **inner_kwargs)",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/orquestra/sdk/_ray/_dag.py", line 147, in __call__',  # noqa: E501
+                    "    return self._fn(*unpacked_args, **unpacked_kwargs)",
+                    '  File "/Users/benjaminmummery/Documents/Projects/orquestra-workflow-sdk/../scratch/scratch.py", line 12, in do_thing',  # noqa: E501
+                    "Exception: Blah",
+                    "2023-03-14 14:32:57,944\tERROR worker.py:399 -- Unhandled error (suppress with 'RAY_IGNORE_UNHANDLED_ERRORS=1'): \x1b[36mray::WorkflowManagementActor.execute_workflow()\x1b[39m (pid=178, ip=10.200.134.19, repr=<ray.workflow.workflow_access.WorkflowManagementActor object at 0x7fc4c4288760>)",  # noqa: E501
+                    "ray.exceptions.RayTaskError: \x1b[36mray::_workflow_task_executor_remote()\x1b[39m (pid=235, ip=10.200.134.19)",  # noqa: E501
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/ray/workflow/task_executor.py", line 115, in _workflow_task_executor_remote',  # noqa: E501
+                    "    return _workflow_task_executor(",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/ray/workflow/task_executor.py", line 84, in _workflow_task_executor',  # noqa: E501
+                    "    raise e",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/ray/workflow/task_executor.py", line 79, in _workflow_task_executor',  # noqa: E501
+                    "    output = func(*args, **kwargs)",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/orquestra/sdk/_ray/_dag.py", line 199, in _ray_remote',  # noqa: E501
+                    "    raise e",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/orquestra/sdk/_ray/_dag.py", line 192, in _ray_remote',  # noqa: E501
+                    "    return wrapped(*inner_args, **inner_kwargs)",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/orquestra/sdk/_ray/_dag.py", line 147, in __call__',  # noqa: E501
+                    "    return self._fn(*unpacked_args, **unpacked_kwargs)",
+                    '  File "/Users/benjaminmummery/Documents/Projects/orquestra-workflow-sdk/../scratch/scratch.py", line 12, in do_thing',  # noqa: E501
+                    "Exception: Blah",
+                    "The above exception was the direct cause of the following exception:",  # noqa: E501
+                    "\x1b[36mray::WorkflowManagementActor.execute_workflow()\x1b[39m (pid=178, ip=10.200.134.19, repr=<ray.workflow.workflow_access.WorkflowManagementActor object at 0x7fc4c4288760>)",  # noqa: E501
+                    '  File "/usr/local/lib/python3.9/concurrent/futures/_base.py", line 439, in result',  # noqa: E501
+                    "    return self.__get_result()",
+                    '  File "/usr/local/lib/python3.9/concurrent/futures/_base.py", line 391, in __get_result',  # noqa: E501
+                    "    raise self._exception",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/ray/workflow/workflow_access.py", line 209, in execute_workflow',  # noqa: E501
+                    "    await executor.run_until_complete(job_id, context, wf_store)",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/ray/workflow/workflow_executor.py", line 109, in run_until_complete',  # noqa: E501
+                    "    await asyncio.gather(",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/ray/workflow/workflow_executor.py", line 356, in _handle_ready_task',  # noqa: E501
+                    "    raise err",
+                    "ray.workflow.exceptions.WorkflowExecutionError: Workflow[id=wf.wf_def.3c5f938] failed during execution.",  # noqa: E501
+                    ":task_name:create_ray_workflow",
+                    ":task_name:_workflow_task_executor_remote",
+                    ":actor_name:WorkflowManagementActor",
+                    "2023-03-14 14:32:57,923\tINFO workflow_executor.py:86 -- Workflow job [id=wf.wf_def.3c5f938] started.",  # noqa: E501
+                    "2023-03-14 14:32:57,938\tERROR workflow_executor.py:306 -- Task status [FAILED] due to an exception raised by the task.\t[wf.wf_def.3c5f938@invocation-0-task-do-thing]",  # noqa: E501
+                    "2023-03-14 14:32:57,942\tERROR workflow_executor.py:352 -- Workflow 'wf.wf_def.3c5f938' failed due to \x1b[36mray::_workflow_task_executor_remote()\x1b[39m (pid=235, ip=10.200.134.19)",  # noqa: E501
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/ray/workflow/task_executor.py", line 115, in _workflow_task_executor_remote',  # noqa: E501
+                    "    return _workflow_task_executor(",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/ray/workflow/task_executor.py", line 84, in _workflow_task_executor',  # noqa: E501
+                    "    raise e",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/ray/workflow/task_executor.py", line 79, in _workflow_task_executor',  # noqa: E501
+                    "    output = func(*args, **kwargs)",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/orquestra/sdk/_ray/_dag.py", line 199, in _ray_remote',  # noqa: E501
+                    "    raise e",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/orquestra/sdk/_ray/_dag.py", line 192, in _ray_remote',  # noqa: E501
+                    "    return wrapped(*inner_args, **inner_kwargs)",
+                    '  File "/home/orquestra/venv/lib/python3.9/site-packages/orquestra/sdk/_ray/_dag.py", line 147, in __call__',  # noqa: E501
+                    "    return self._fn(*unpacked_args, **unpacked_kwargs)",
+                    '  File "/Users/benjaminmummery/Documents/Projects/orquestra-workflow-sdk/../scratch/scratch.py", line 12, in do_thing',  # noqa: E501
+                    "Exception: Blah",
+                    ":actor_name:WorkflowManagementActor",
+                ]
+
+            @staticmethod
             def test_params_encoding(
                 endpoint_mocker, client: DriverClient, workflow_run_id: str
             ):
@@ -1609,7 +1689,8 @@ class TestClient:
                 Verifies that params are correctly sent to the server.
                 """
                 endpoint_mocker(
-                    json=resp_mocks.make_get_wf_run_logs_response(),
+                    # json=resp_mocks.make_get_wf_run_logs_response(),
+                    body=resp_mocks.make_get_wf_run_logs_response_with_content(),
                     match=[
                         responses.matchers.query_param_matcher(
                             {"workflowRunId": workflow_run_id}
@@ -1648,11 +1729,28 @@ class TestClient:
                     _ = client.get_workflow_run_logs(workflow_run_id)
 
             @staticmethod
+            def test_zlib_error(
+                endpoint_mocker, client: DriverClient, workflow_run_id: str
+            ):
+                endpoint_mocker(
+                    body=b"invalid bytes",
+                    match=[
+                        responses.matchers.query_param_matcher(
+                            {"workflowRunId": workflow_run_id}
+                        )
+                    ],
+                )
+
+                with pytest.raises(_exceptions.WorkflowRunLogsNotReadable):
+                    _ = client.get_workflow_run_logs(workflow_run_id)
+
+            @staticmethod
             def test_sets_auth(
                 endpoint_mocker, client: DriverClient, token: str, workflow_run_id: str
             ):
                 endpoint_mocker(
-                    json=resp_mocks.make_get_wf_run_logs_response(),
+                    # json=resp_mocks.make_get_wf_run_logs_response(),
+                    body=resp_mocks.make_get_wf_run_logs_response_with_content(),
                     match=[
                         responses.matchers.header_matcher(
                             {"Authorization": f"Bearer {token}"}